refactored code to be simple prior to migrating to Maui message center
