+++ src/apps/NotificationHub.Maui/Platforms/Android/Helpers/Constants.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NotificationHub.Maui.Platforms.Android.Helpers
+{
+    internal class Constants
+    {
+        public const string NOTIFICATION_CHANNEL_ID = "notification_hub_channel";
+    }
+}
