+++ src/apps/NotificationHub.Maui/Platforms/Android/MainActivity.cs
@@ -18,8 +18,6 @@ namespace NotificationHub.Maui.Platforms.Android
     [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
     public class MainActivity : MauiAppCompatActivity
     {
-        public static readonly string CHANNEL_ID = "notification_hub_channel";
-
         public MainActivity()
         {
